Add optional `queue` to handler
For async servers, there has to be a synchronization point
for done callbacks. This is what the `queue` is for.

If the handler doesn't dispatch to other queues but stays
within the calling one, it doesn't need to call the done
callbacks via

    queue.async { done() }

but can do a straight

    done()

But if the handler does perform its work async, it need
to use the given queue to call the done, e.g.:

    func workHard(req: HTTPRequest, res: HTTPResponseWriter, queue: DispatchQueue?) {
        DispatchQueue.background.async {
            // do hard work
            queue.async {
               res.writeHeader(status: .ok)
            }
        }
    }

etc.

We are using `DispatchQueue` here. Another option is to add a
non-optional

  HTTPSynchronizationContext protocol

(adapted by DispatchQueue or a uv or sync variant)